+++ TODO.R
@@ -1,11 +1,3 @@
-# FIXME screenshot on quarto-blog so that, it only show image.png
-# Use with quarto, Rmd (source mode) with
-# ![](posts/openxlsx2-1-0){fig-alt="" width="70%"}
-# ℹ Consider using a more precise name
-# rename_file('posts/openxlsx2-1-0/image-01.png', 'posts/openxlsx2-1-0/better-name.png')
-# ℹ See reuseme::rename_file for details.
-# TODO screenshot increments by 2 image-01, then image-03
-# TODO screenshot addin to insert the code directly in the qmd doc.
-# TODO screenshot overwrites.
-# TODO todo could guess where it should be written. Maybe om top of an R file?
-# TODO make the screenshot behaviour different when vignettes vs articles: vignettes should place it in man/figures, while articles could put it in vignettes/articles file.
+# TODO [todo] todo could guess where it should be written. Maybe om top of an R file?
+# TODO [screenshot] make the behaviour different when vignettes vs articles: vignettes should place it in man/figures, while articles could put it in vignettes/articles file.
+# TODO [screenshot] RStudio addin to insert the code directly in the qmd doc.